UserListsの表示ロジックをUserLists.tsxに記述へ変更、movieListDataの構造変更
UserListsの表示ロジックをuser/page.tsxに記述していたがUserLists.tsxに変更

movieListDataの構造をuid の階層を変更
movieListData→uid→movieId
から
movieListData→movieId
に変更
