+++ src/app/user/page.tsx
@@ -1,19 +1,13 @@
 'use client'
 
-import { useEffect, useState } from 'react'
+import { useEffect } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
-import { toast } from 'react-toastify'
 
-import {
-  fetchUserLists,
-  // updateComment,
-} from '@/lib/features/movies/moviesSlice'
+import { fetchUserLists } from '@/lib/features/movies/moviesSlice'
 import { AppDispatch, RootState } from '@/lib/store'
-import { toastConfig } from '@/lib/toastConfig'
 
 import Profile from '../../components/Profile'
 import UserLists from '../../components/UserLists'
-import { ListType } from '../../types/Lists'
 import { User } from '../../types/User'
 import Loading from '../loading'
 
@@ -21,44 +15,11 @@ export default function Home() {
   const dispatch: AppDispatch = useDispatch()
   const user: User | null = useSelector((state: RootState) => state.auth.user)
 
-  const movies = useSelector((state: RootState) =>
-    user ? state.movies.movieListData[user.uid] : undefined,
-  )
-
-  const [edittingMovieId, setEdittingMovieId] = useState<string | null>(null)
-  const [inputedComment, setInputedComment] = useState<string>('')
-
   useEffect(() => {
     if (user) {
       dispatch(fetchUserLists(user.uid))
     }
   }, [user, dispatch])
-
-  const toggleEditMode = (movieId: string, comment: string | undefined) => {
-    setEdittingMovieId((prevId) => (prevId === movieId ? null : movieId))
-    setInputedComment(comment || '')
-  }
-
-  const confirmEdit = (movieId: string, uid: string) => {
-    setEdittingMovieId(null)
-
-    // dispatch(updateComment({ movieId, uid, comment: inputedComment }))
-    toast.success('コメントを編集しました', toastConfig)
-  }
-
-  const cancelEdit = () => {
-    setEdittingMovieId(null)
-    if (window.confirm('編集をキャンセルしますか？')) {
-      setInputedComment('')
-      toast.error('編集をキャンセルしました', toastConfig)
-    } else {
-      return
-    }
-  }
-
-  const handleOnChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
-    setInputedComment(e.target.value)
-  }
 
   if (!user) {
     return <Loading />
@@ -71,17 +32,7 @@ export default function Home() {
       </h1>
 
       <Profile />
-
-      {movies && (
-        <UserLists
-          edittingMovieId={edittingMovieId}
-          inputedComment={inputedComment}
-          handleOnChange={handleOnChange}
-          confirmEdit={confirmEdit}
-          cancelEdit={cancelEdit}
-          toggleEditMode={toggleEditMode}
-        />
-      )}
+      <UserLists />
     </div>
   )
 }
