+++ src/types/Lists.ts
@@ -13,5 +13,5 @@ export interface MovieItem {
   movieId: string
   addedAt: string
   comment?: string
-  watchedAt?: Timestamp | null
+  watchedAt?: string | null
 }
